refactor(api): move oidc-authentication-service.js to src/authentication
