+++ api/lib/domain/services/authentication/paysdelaloire-oidc-authentication-service.js
@@ -1,6 +1,6 @@
+import { OidcAuthenticationService } from '../../../../src/authentication/domain/services/oidc-authentication-service.js';
 import { config } from '../../../../src/shared/config.js';
 import { PAYSDELALOIRE } from '../../constants/oidc-identity-providers.js';
-import { OidcAuthenticationService } from './oidc-authentication-service.js';
 
 const configKey = PAYSDELALOIRE.configKey;
