+++ api/lib/domain/services/authentication/pole-emploi-oidc-authentication-service.js
@@ -1,12 +1,12 @@
 import { randomUUID } from 'crypto';
 import dayjs from 'dayjs';
 
+import { OidcAuthenticationService } from '../../../../src/authentication/domain/services/oidc-authentication-service.js';
 import { config } from '../../../config.js';
 import { DomainTransaction } from '../../../infrastructure/DomainTransaction.js';
 import { temporaryStorage } from '../../../infrastructure/temporary-storage/index.js';
 import { POLE_EMPLOI } from '../../constants/oidc-identity-providers.js';
-import { AuthenticationMethod } from '../../models/AuthenticationMethod.js';
-import { OidcAuthenticationService } from './oidc-authentication-service.js';
+import { AuthenticationMethod } from '../../models/index.js';
 
 const configKey = POLE_EMPLOI.configKey;
 const logoutUrlTemporaryStorage = temporaryStorage.withPrefix('logout-url:');
