+++ api/tests/authentication/unit/domain/services/oidc-authentication-service_test.js
@@ -5,12 +5,14 @@ import { config as settings } from '../../../../../lib/config.js';
 import { OIDC_ERRORS } from '../../../../../lib/domain/constants.js';
 import * as OidcIdentityProviders from '../../../../../lib/domain/constants/oidc-identity-providers.js';
 import { OidcMissingFieldsError } from '../../../../../lib/domain/errors.js';
-import { AuthenticationMethod } from '../../../../../lib/domain/models/AuthenticationMethod.js';
-import { AuthenticationSessionContent } from '../../../../../lib/domain/models/AuthenticationSessionContent.js';
-import { UserToCreate } from '../../../../../lib/domain/models/UserToCreate.js';
-import { OidcAuthenticationService } from '../../../../../lib/domain/services/authentication/oidc-authentication-service.js';
-import { DomainTransaction } from '../../../../../lib/infrastructure/DomainTransaction.js';
+import {
+  AuthenticationMethod,
+  AuthenticationSessionContent,
+  UserToCreate,
+} from '../../../../../lib/domain/models/index.js';
 import { monitoringTools } from '../../../../../lib/infrastructure/monitoring-tools.js';
+import { OidcAuthenticationService } from '../../../../../src/authentication/domain/services/oidc-authentication-service.js';
+import { DomainTransaction } from '../../../../../src/shared/domain/DomainTransaction.js';
 import { OidcError } from '../../../../../src/shared/domain/errors.js';
 import { logger } from '../../../../../src/shared/infrastructure/utils/logger.js';
 import { catchErr, catchErrSync, expect, sinon } from '../../../../test-helper.js';
