+++ api/tests/unit/setup-oidc-authentication-service-registry_test.js
@@ -1,6 +1,6 @@
 import { setupOidcAuthenticationServiceRegistry } from '../../config/setup-oidc-authentication-service-registry.js';
 import { oidcAuthenticationServiceRegistry } from '../../lib/domain/services/authentication/authentication-service-registry.js';
-import { OidcAuthenticationService } from '../../lib/domain/services/authentication/oidc-authentication-service.js';
+import { OidcAuthenticationService } from '../../src/authentication/domain/services/oidc-authentication-service.js';
 import { expect, sinon } from '../test-helper.js';
 
 describe('Unit | Config | setup-oidc-authentication-service-registry', function () {
